requirements.txt
fastapi
uvicorn[standard]

torch
transformers
accelerate

pillow
opencv-python
supervision==0.19.0

chromadb
pymongo

numpy
python-dotenv
python-multipart
